refactor: split code for better testing
